+++ src/Resources/contao/languages/en/tl_module.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * Contao Mailinglisten Bundle.
+ *
+ * @license LGPL-3.0-or-later
+ */
+
+// Fields
+$GLOBALS['TL_LANG']['tl_module']['mlListe'] = array('Mailing list', 'The list visitors can subscribe to through this form. Unpublished lists can be selected, but the form only appears once the list is published.');
+$GLOBALS['TL_LANG']['tl_module']['mlEinwilligung'] = array('Privacy notice', 'Shown as a checkbox below the form, which then has to be ticked. HTML is allowed, e.g. for a link to the privacy policy. Leave empty to omit the checkbox.');